refactor: remove DB import from TsUtils
