+++ src/lib/tabs/tabOptimizer/optimizerForm/components/CharacterSelect.tsx
@@ -1,7 +1,7 @@
 import { Button, Card, Flex, Input, InputRef, Modal, Select } from 'antd'
 import { Assets } from 'lib/rendering/assets'
+import { generateCharacterOptions } from 'lib/rendering/optionGenerator'
 import { CardGridItemContent, generateElementTags, generatePathTags, SegmentedFilterRow } from 'lib/tabs/tabOptimizer/optimizerForm/components/CardSelectModalComponents'
-import { TsUtils } from 'lib/utils/TsUtils'
 import { Utils } from 'lib/utils/utils'
 import * as React from 'react'
 import { ReactElement, useEffect, useMemo, useRef, useState } from 'react'
@@ -40,7 +40,7 @@ const CharacterSelect: React.FC<CharacterSelectProps> = ({ value, onChange, sele
   const { t } = useTranslation('modals', { keyPrefix: 'CharacterSelect' })
   const [open, setOpen] = useState(false)
   const [currentFilters, setCurrentFilters] = useState(Utils.clone(defaultFilters))
-  const characterOptions = useMemo(() => TsUtils.generateCharacterOptions(), [t])
+  const characterOptions = useMemo(() => generateCharacterOptions(), [t])
   const [selected, setSelected] = useState<Map<string, boolean>>(new Map())
   const excludedRelicPotentialCharacters = window.store((s) => s.excludedRelicPotentialCharacters)
 
@@ -240,11 +240,11 @@ const CharacterSelect: React.FC<CharacterSelectProps> = ({ value, onChange, sele
                         height: `${parentH}px`,
                       },
                       ...(selected.get(option.id)
-                        ? {
-                          opacity: 0.25,
-                          background: 'grey',
-                        }
-                        : {}
+                          ? {
+                            opacity: 0.25,
+                            background: 'grey',
+                          }
+                          : {}
                       ),
                     }}
                     styles={{ body: { padding: 1 } }}
